Add option to not use Web Worker for API simulation, noticed performance boost
